+++ locals.tf
@@ -24,4 +24,6 @@ locals {
 
   backup_resource_group_name = var.backup_policy_id != null ? split("/", var.backup_policy_id)[4] : null
   backup_recovery_vault_name = var.backup_policy_id != null ? split("/", var.backup_policy_id)[8] : null
+
+  key_vault_name = split("/", var.key_vault_id)[8]
 }